requirements.txt
azure-functions
requests
pillow
pytesseract
twilio
python-dotenv

